Fix regex for attribute names to allow unquoted substitutions
The following code snippet is currently valid Nix code:

```nix
let
  name = "foo";
in
{
  ${name} = "bar";
}
```

Until now the `attribute-or-function` matcher which is always used when
it's not clear if the current expression is an attribute set or a
function with an attribute set matching. The regex which should match an
attribute set inside the matcher didn't cover the case of such a
quoting.

Addresses #8 (but doesn't fix it yet)